Revert "VZ-3747 Part 2 - Uptake dashboard changes to add cluster drop down, remove old managed_cluster label, update tests (#1919)"
This reverts commit 4ca17b052676d9f65b5c4938ab695992eb1e1099.
