+++ tests/e2e/multicluster/examples/helidon-deprecated/helidon_example_test.go
@@ -163,9 +163,8 @@ var _ = Describe("Multi-cluster verify hello-helidon", func() {
 	// THEN expect Prometheus metrics for the app to exist in Prometheus on the admin cluster
 	Context("Metrics", func() {
 		It("Verify Prometheus metrics exist on admin cluster", func() {
-			clusterNameMetricsLabel, _ := pkg.GetClusterNameMetricLabel()
 			Eventually(func() bool {
-				var m = map[string]string{clusterNameMetricsLabel: clusterName}
+				var m = map[string]string{"managed_cluster": clusterName}
 				return pkg.MetricsExistInCluster("base_jvm_uptime_seconds", m, adminKubeconfig)
 			}, longWaitTimeout, longPollingInterval).Should(BeTrue())
 		})
