+++ tests/e2e/multicluster/examples/todo-list/todo_list_example_test.go
@@ -206,31 +206,28 @@ var _ = Describe("Multi-cluster verify todo-list", func() {
 	Context("Prometheus Metrics", func() {
 
 		It("Verify scrape_duration_seconds metrics exist for managed cluster", func() {
-			clusterNameMetricsLabel, _ := pkg.GetClusterNameMetricLabel()
 			Eventually(func() bool {
 				m := make(map[string]string)
 				m["namespace"] = testNamespace
-				m[clusterNameMetricsLabel] = clusterName
+				m["managed_cluster"] = clusterName
 				return pkg.MetricsExistInCluster("scrape_duration_seconds", m, adminKubeconfig)
 			}, longWaitTimeout, longPollingInterval).Should(BeTrue(), "Expected to find base_jvm_uptime_seconds metric")
 		})
 
 		It("Verify DNE scrape_duration_seconds metrics does not exist for managed cluster", func() {
-			clusterNameMetricsLabel, _ := pkg.GetClusterNameMetricLabel()
 			Eventually(func() bool {
 				m := make(map[string]string)
 				m["namespace"] = testNamespace
-				m[clusterNameMetricsLabel] = "DNE"
+				m["managed_cluster"] = "DNE"
 				return pkg.MetricsExistInCluster("scrape_duration_seconds", m, adminKubeconfig)
 			}, longWaitTimeout, longPollingInterval).Should(BeFalse(), "Not expected to find base_jvm_uptime_seconds metric")
 		})
 
 		It("Verify container_cpu_cfs_periods_total metrics exist for managed cluster", func() {
-			clusterNameMetricsLabel, _ := pkg.GetClusterNameMetricLabel()
 			Eventually(func() bool {
 				m := make(map[string]string)
 				m["namespace"] = testNamespace
-				m[clusterNameMetricsLabel] = clusterName
+				m["managed_cluster"] = clusterName
 				return pkg.MetricsExistInCluster("container_cpu_cfs_periods_total", m, adminKubeconfig)
 			}, longWaitTimeout, longPollingInterval).Should(BeTrue(), "Expected to find container_cpu_cfs_periods_total metric")
 		})
